+++ src/main/java/org/example/fanzip/cart/controller/CartController.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.RequestHeader;
-import org.example.fanzip.auth.jwt.JwtProcessor;
 
 import javax.servlet.http.HttpServletRequest;
