+++ src/main/java/org/example/fanzip/membership/controller/MembershipController.java
@@ -1,10 +1,10 @@
 package org.example.fanzip.membership.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.fanzip.auth.jwt.JwtProcessor;
 import org.example.fanzip.membership.dto.MembershipSubscribeRequestDTO;
 import org.example.fanzip.membership.dto.MembershipSubscribeResponseDTO;
 import org.example.fanzip.membership.service.MembershipService;
+import org.example.fanzip.security.JwtProcessor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,7 +22,7 @@ public class MembershipController {
             @RequestBody MembershipSubscribeRequestDTO requestDTO
     ) {
         String token = authorizationHeader.substring(7);
-        Long userId = jwtProcessor.getUserId(token);
+        Long userId = jwtProcessor.getUserIdFromToken(token);
 
         MembershipSubscribeResponseDTO responseDTO =
                 membershipService.subscribe(requestDTO, userId);
